fix: fixes issue with player not getting destroyed from memory
